FSR-980 move server files to lib folder
